Module alchemy::distro

Top Level Module File All of the various available public modules that can be used are listed here All common functions related to distributions as well as distribution specific functions are available in the distro module

Enums

Distro

Functions

which_distro
Returns what Distribution the user is using If no possible match is found return a None