Module alchemy::kernel::vga

This module contains a VGA driver which makes use of a Mutex provided by the spin crate (spinlock) and an AtomicPtr to enable access to an instance of the Writer structure statically.

Structs

ColorPair
A pair of Colors which represents foreground/background. Note that both colors are packed into a single u8 to save space. A VGAChar is 16-bit and thus the color forms the latter byte.

VGAChar
This structure represents a single character on the VGA display and its formatting options (only foreground and background color at present). Each character is two bytes, with the first byte containing the ASCII character code and the latter byte containing the formatting options.

Writer
Writer is a representation of VGA state, it contains positioning information for the current row and column, as well as formatting options and an AtomicPtr to a Buffer.

Enums

Color
Enumerative type which is used for foreground and backgroun colors alike.

Constants

HEIGHT
Height of the VGA output in characters.

WIDTH
Width of the VGA output in characters.

Statics

VGA
A static VGA which makes use of spin::Mutex to provide locking. It is this static structure which makes println! et al. possible.