Struct alchemy::kernel::vga::Writer

pub struct Writer {
    pub color: ColorPair,
    // some fields omitted
}

Writer is a representation of VGA state, it contains positioning information for the current row and column, as well as formatting options and an AtomicPtr to a Buffer.

Fields
color: ColorPair
Methods

impl Writer

const fn new(color: ColorPair) -> Writer

Creates a new Writer from a ColorPair. Note that the buffer pointer is hardcoded.

fn write_byte(&mut self, byte: u8)

Writes a single byte to the screen, and takes care to handle the edge of the screen and newlines appropriately by updating the structure's internal state.