Struct alchemy::models::Package

pub struct Package {
    pub id: i32,
    pub arch: String,
    pub aur: String,
    pub void: String,
    pub ubuntu: String,
    pub ubuntu_dev: String,
}

Struct for a row in the Alchemist package table

Fields

id

Row Number makes it easy for migration

arch

Arch Linux Package

aur

Arch Linux AUR Package

void

Void Linux Package

ubuntu

Ubuntu Binary Packages

ubuntu_dev

Ubuntu Development Header Packages

Methods

impl Package

fn empty() -> Package

Returns a Package with no data for instances of error handling when connecting to the db and finding nothing